Added the Delete task method. I included a few methods related to the delete task method as well.
I added Comments for all methods and made notes where needed.

I also tested it with multiple scenarios and everything seems to be working as intended.

I also found that when creating a task it would crash. This was due to the schedule array never being initialized. So i fixed that and all seems to be working